create Neon/TextField component
